requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-cleanup==8.0.0
django-crispy-forms==2.0
django-humanize==0.1.2
django-jalali-date==1.0.2
django-widget-tweaks==1.4.12
humanize==4.7.0
idna==3.4
jdatetime==4.1.1
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.6
